+++ R/buckets.R
@@ -21,8 +21,10 @@ gcs_global_bucket <- function(bucket){
     bucket <- bucket$name
   }
 
-  stopifnot(inherits(bucket, "character"),
-            length(bucket) == 1)
+  assertthat::assert_that(
+    is.character(bucket),
+    is.unit(bucket)
+  )
 
   .gcs_env$bucket <- bucket
   message("Set default bucket name to '", bucket,"'")
@@ -57,7 +59,7 @@ gcs_get_global_bucket <- function(){
 #' List the buckets your projectId has access to
 #'
 #' @param projectId Project containing buckets to list
-#' @param prefix Filter results to names begnning with this prefix
+#' @param prefix Filter results to names beginning with this prefix
 #' @param projection Properties to return. Default noAcl omits acl properties
 #' @param maxResults Max number of results
 #' @param detail Set level of detail
@@ -71,7 +73,18 @@ gcs_get_global_bucket <- function(){
 #'   \item \code{full} - as above plus: id, selfLink, projectNumber, timeCreated, metageneration, etag
 #'  }
 #'
-#' @return data.frame of buckets
+#' @return \code{data.frame} of buckets
+#'
+#' @examples
+#'
+#' \dontrun{
+#'
+#' buckets <- gcs_list_buckets("your-project")
+#'
+#' ## use the name of the bucket to get more meta data
+#' bucket_meta <- gcs_get_bucket(buckets$name[[1]])
+#'
+#' }
 #'
 #' @family bucket functions
 #' @export
@@ -84,9 +97,10 @@ gcs_list_buckets <- function(projectId,
   projection <- match.arg(projection)
   detail <- match.arg(detail)
 
-  testthat::expect_is(projectId, "character")
-  testthat::expect_is(prefix, "character")
-  testthat::expect_is(maxResults, "numeric")
+  assertthat::assert_that(is.character(projectId),
+                          is.unit(projectId),
+                          is.character(prefix),
+                          is.numeric(maxResults))
 
   parse_lb <- function(x){
     x <- x$items
@@ -114,6 +128,7 @@ gcs_list_buckets <- function(projectId,
 
   out[,out_names]
 
+
 }
 
 #' Get bucket info
@@ -124,6 +139,17 @@ gcs_list_buckets <- function(projectId,
 #' @param ifMetagenerationMatch Return only if metageneration matches
 #' @param ifMetagenerationNotMatch Return only if metageneration does not match
 #' @param projection Properties to return. Default noAcl omits acl properties
+#'
+#' @examples
+#'
+#' \dontrun{
+#'
+#' buckets <- gcs_list_buckets("your-project")
+#'
+#' ## use the name of the bucket to get more meta data
+#' bucket_meta <- gcs_get_bucket(buckets$name[[1]])
+#'
+#' }
 #'
 #' @return A bucket resource object
 #' @family bucket functions
@@ -139,8 +165,9 @@ gcs_get_bucket <- function(bucket = gcs_get_global_bucket(),
     bucket <- bucket$name
   }
 
-  testthat::expect_is(bucket, "character")
-  testthat::expect_is(projection, "character")
+  assertthat::assert_that(is.character(bucket),
+                          is.unit(bucket),
+                          is.character(projection))
 
   pars_args <- list(ifMetagenerationMatch=ifMetagenerationMatch,
                     ifMetagenerationNotMatch=ifMetagenerationNotMatch,
@@ -183,16 +210,17 @@ gcs_create_bucket <-
            storageClass = c("STANDARD",
                             "NEARLINE",
                             "DURABLE_REDUCED_AVAILABILITY"),
-           predefinedAcl = c("authenticatedRead",
+           predefinedAcl = c("projectPrivate",
+                             "authenticatedRead",
                              "private",
-                             "projectPrivate",
                              "publicRead",
                              "publicReadWrite"),
-           predefinedDefaultObjectAcl = c("authenticatedRead",
+           predefinedDefaultObjectAcl = c("bucketOwnerFullControl",
+                                          "bucketOwnerRead",
+                                          "authenticatedRead",
                                           "private",
                                           "projectPrivate",
-                                          "publicRead",
-                                          "publicReadWrite"),
+                                          "publicRead"),
            projection = c("noAcl","full"),
            versioning = FALSE,
            lifecycle = NULL){
@@ -202,11 +230,16 @@ gcs_create_bucket <-
   storageClass  <- match.arg(storageClass)
   predefinedDefaultObjectAcl <- match.arg(predefinedDefaultObjectAcl)
 
-  testthat::expect_is(projectId, "character")
-  testthat::expect_is(name, "character")
-  testthat::expect_is(location, "character")
-  testthat::expect_is(projection, "character")
-  testthat::expect_type(versioning, "logical")
+  assertthat::assert_that(
+    is.character(projectId),
+    is.character(name),
+    is.unit(projectId),
+    is.unit(name),
+    is.character(location),
+    is.unit(location),
+    is.character(projection),
+    is.logical(versioning)
+  )
 
   pars_args <- list(project = projectId,
                     predefinedAcl = predefinedAcl,
@@ -254,7 +287,6 @@ gcs_create_bucket <-
 #' @param ifMetagenerationMatch Delete only if metageneration matches
 #' @param ifMetagenerationNotMatch Delete only if metageneration does not match
 #'
-#' @import testthat
 #' @family bucket functions
 #' @export
 gcs_delete_bucket <- function(bucket,
@@ -265,7 +297,8 @@ gcs_delete_bucket <- function(bucket,
     bucket <- bucket$name
   }
 
-  testthat::expect_is(bucket, "character")
+  assertthat::assert_that(is.character(bucket),
+                          is.unit(bucket))
 
   pars_args <- list(ifMetagenerationMatch=ifMetagenerationMatch,
                     ifMetagenerationNotMatch=ifMetagenerationNotMatch)
@@ -299,7 +332,7 @@ gcs_delete_bucket <- function(bucket,
 #' @param age Age in days before objects are deleted
 #' @param createdBefore Deletes all objects before this date
 #' @param numNewerVersions Deletes all newer versions of this object
-#' @param isLive If TRUE deletes all live objects, if FALSE delets all archived versions
+#' @param isLive If TRUE deletes all live objects, if FALSE deletes all archived versions
 #'
 #' \code{numNewerVersions} and \code{isLive} works only for buckets with object versioning
 #'
@@ -318,7 +351,7 @@ gcs_create_lifecycle <- function(age = NULL,
                                  isLive = NULL){
 
   if(!is.null(age)){
-    testthat::expect_is(age, "numeric")
+    assertthat::assert_that(is.numeric(age))
   }
 
 
@@ -327,15 +360,15 @@ gcs_create_lifecycle <- function(age = NULL,
     if(is.na(createdBefore)){
       stop("Problem with createdBefore date, converted to ", createdBefore)
     }
-    testthat::expect_type(createdBefore, "character")
+    assertthat::assert_that(is.character(createdBefore))
   }
 
   if(!is.null(numNewerVersions)){
-    testthat::expect_is(numNewerVersions, "numeric")
+    assertthat::assert_that(is.numeric(numNewerVersions))
   }
 
   if(!is.null(isLive)){
-    testthat::expect_is(isLive, "logical")
+    assertthat::assert_that(is.logical(isLive))
   }
 
   rule <- list(
@@ -362,10 +395,13 @@ make_lifecycle_list <- function(lifecycle){
 
   if(!is.null(lifecycle)){
 
-    testthat::expect_type(lifecycle, "list")
-    testthat::expect_true(all(vapply(lifecycle,
-                                     function(x) inherits(x, "gcs_lifecycle"),
-                                     logical(1))))
+    assertthat::assert_that(
+      is.list(lifecycle),
+      all(vapply(lifecycle,
+                 function(x) inherits(x, "gcs_lifecycle"),
+                 logical(1)))
+      )
+
     out <- list(
       list(
         rule = lifecycle