+++ R/utilities.R
@@ -1,3 +1,7 @@
+is.unit <- function(x){
+  length(x) == 1L
+}
+
 #' if argument is NULL, no line output
 #'
 #' @keywords internal
@@ -7,6 +11,12 @@ cat0 <- function(prefix = "", x){
   }
 }
 
+#' Javascript time to R time
+#'
+#' @keywords internal
+js_to_posix <- function(x){
+  as.POSIXct(as.numeric(x) / 1000, origin = "1970-01-01")
+}
 
 #' taken from utils:::format.object_size
 #'